Metafile multi-line support. Add a test for this.
The Metafile class now saves to disk using yaml.dump. The unit
test for banners (or anything) spanning multiple lines is been
added as test/metafile_test_banner.py.

To format a banner, start and end lines with an escape character
if they would otherwise start or end with a space, so yaml.dump
won't strip the space. Put four spaces before the first character
in a line.