Add error check for Metafile.get() on an undefined tag + test
If Metafile.get(tag) is called on a tag which does not exist in
Metafile.data, a KeyError is caught and None is returned. In this
way, the syntax "Metafile.get(tag) or default" can be used to load
a default value for a tag if it has not already been defined.